Add Web/Desktop specific plugin capabilities; misc fixes
