+++ src/plugins/arRPC.web.tsx
@@ -48,7 +48,6 @@ export default definePlugin({
     name: "WebRichPresence (arRPC)",
     description: "Client plugin for arRPC to enable RPC on Discord Web (experimental)",
     authors: [Devs.Ducko],
-    target: "WEB",
 
     settingsAboutComponent: () => (
         <>
@@ -60,6 +59,9 @@ export default definePlugin({
     ),
 
     async start() {
+        // ArmCord comes with its own arRPC implementation, so this plugin just confuses users
+        if ("armcord" in window) return;
+
         if (ws) ws.close();
         ws = new WebSocket("ws://127.0.0.1:1337"); // try to open WebSocket
