+++ src/plugins/noRPC.desktop.ts
@@ -16,27 +16,22 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+import { migratePluginSettings } from "@api/settings";
 import { Devs } from "@utils/constants";
 import definePlugin from "@utils/types";
 
+migratePluginSettings("NoRPC", "No RPC");
 export default definePlugin({
-    name: "NoSystemBadge",
-    description: "Disables the taskbar and system tray unread count badge.",
-    authors: [Devs.rushii],
-    target: "DESKTOP",
+    name: "NoRPC",
+    description: "Disables Discord's RPC server.",
+    authors: [Devs.Cyn],
     patches: [
         {
-            find: "setSystemTrayApplications:function",
-            replacement: [
-                {
-                    match: /setBadge:function.+?},/,
-                    replace: "setBadge:function(){},"
-                },
-                {
-                    match: /setSystemTrayIcon:function.+?},/,
-                    replace: "setSystemTrayIcon:function(){},"
-                }
-            ]
-        }
-    ]
+            find: '.ensureModule("discord_rpc")',
+            replacement: {
+                match: /\.ensureModule\("discord_rpc"\)\.then\(\(.+?\)\)}/,
+                replace: '.ensureModule("discord_rpc")}',
+            },
+        },
+    ],
 });
